chore: refectoring add image

+++ src/services/apiCabins.ts
@@ -1,4 +1,4 @@
-import { CabinType } from "../types/CabinsType";
+import { CabinFormData } from "../types/CabinFornData";
 import supabase, { supabaseUrl } from "./supabase";
 import { PostgrestError } from "@supabase/supabase-js";
 
@@ -14,42 +14,53 @@ export async function getCabins() {
 }
 
 export async function createEditCabin(
-  newCabin: CabinType,
-  id: CabinType["id"] | null
+  newCabin: CabinFormData,
+  id: number | null
 ) {
-  //newCabin.image?.startsWith?.(supabaseUrl);
- // const hasImagePath = "xxxx"
- // const imageName = `${Math.random()}-${newCabin.image}`.replaceAll("/", "");
- // const imagePath = hasImagePath
-//    ? newCabin.image
- //   : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+ console.log("newCabin - xxxx", newCabin);
+  const hasImagePath =
+    typeof newCabin.image === "string" &&
+    newCabin.image.startsWith(supabaseUrl);
+
+/*   const imageName = `${Date.now()}-${Math.floor(Math.random() * 10000)}-${
+    newCabin.image instanceof FileList ? newCabin.image[0].name : ""
+  }`.replaceAll("/", ""); */
+
+  const imageName = newCabin.image instanceof FileList ? newCabin.image[0].name : "xxx"
+
+
+  console.log("imageName", imageName);
+
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
   // 1. Create/edit cabin
-  let { data, error }: { data: CabinType | null; error: PostgrestError | null } = { data: null, error: null };
-
-
+  let data;
+  let error: PostgrestError | null;
 
   // A) CREATE
   if (!id) {
-    const response = await supabase
+    const result = await supabase
       .from("cabins")
-      .insert([{ ...newCabin}]);
+      .insert([{ ...newCabin, image: imagePath }])
+      .select()
+      .single();
 
-    data = response.data;
-    error = response.error;
+    data = result.data;
+    error = result.error;
   }
-
   // B) EDIT
-  if (id) {
-    const response = await supabase
+  else {
+    const result = await supabase
       .from("cabins")
       .update({ ...newCabin, image: imagePath })
       .eq("id", id)
       .select()
       .single();
 
-    data = response.data;
-    error = response.error;
+    data = result.data;
+    error = result.error;
   }
 
   if (error) {
@@ -58,23 +69,42 @@ export async function createEditCabin(
   }
 
   // 2. Upload image
-/*   if (hasImagePath) return data;
+  if (hasImagePath) return data;
 
+// Verificar se newCabin.image é um arquivo válido antes de fazer upload
+  if (newCabin.image instanceof File) {
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin IF there was an error uplaoding image
+  // 3.ete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
       "Cabin image could not be uploaded and the cabin was not created"
     );
-  } */
+  }
+} else if (newCabin.image instanceof FileList && newCabin.image.length > 0) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image[0]);
+    
+    // 3. Delete the cabi there was an error uplaoding image
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.error(storageError);
+      throw new Error(
+        "Cabin image could not be uploaded and the cabin was not created"
+      );
+    }
+  } else {
+    console.warn("No valid image file to upload");
+  }
 
   return data;
 }
+
 export async function deleteCabin(id: number) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
